FR_716_Smoke_tests
- added some test for notification
